Managing state for intinereiesInfo
